\name{rvwCpp-package}
\alias{rvwCpp-package}
\alias{rvwCpp}
\docType{package}
\title{
  Test implementation of R interface for Vowpal Wabbit
}
\description{
  Basic test implementation of R interface for Vowpal Wabbit using Rcpp and libvw for GSoC 2018.
  This test initial package implements binary classification.
  Right now it uses standart example parser from libvw which is not really faster then command line rvw.
}

\author{
Ivan Pavlov

Maintainer: Ivan Pavlov <pavlov.ivan095@gmail.com>
}
